+++ tests/irdl/test_operation_definition.py
@@ -16,8 +16,8 @@ from xdsl.dialects.builtin import (
     i32,
     i64,
 )
-from xdsl.dialects.test import TestType
-from xdsl.ir import Block, Region
+from xdsl.dialects.test import TestTermOp, TestType
+from xdsl.ir import Attribute, Block, OpResult, Region, SSAValue, SSAValues
 from xdsl.irdl import (
     AnyAttr,
     AnyInt,
@@ -26,6 +26,7 @@ from xdsl.irdl import (
     AttrSizedOperandSegments,
     AttrSizedRegionSegments,
     AttrSizedResultSegments,
+    AttrSizedSuccessorSegments,
     BaseAttr,
     EqAttrConstraint,
     IntVarConstraint,
@@ -54,6 +55,7 @@ from xdsl.irdl import (
     var_operand_def,
     var_region_def,
     var_result_def,
+    var_successor_def,
 )
 from xdsl.parser import Parser
 from xdsl.traits import NoTerminator
@@ -618,6 +620,100 @@ def test_undefined_property():
         + " Use the dictionary attribute to add arbitrary information to the operation.",
     ):
         op.verify()
+
+
+def test_op_accessor_assignment():
+    """Test that operation constructs cannot be assigned via an op's named constructs."""
+    val1, val2 = create_ssa_value(i32), create_ssa_value(i32)
+    op1 = TestTermOp.create(operands=[val1, val2])
+    op1.verify()
+
+    new_operands: SSAValues[SSAValue[Attribute]] = SSAValues((val1, val2))
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.ops = new_operands
+
+    new_results: SSAValues[OpResult[Attribute]] = SSAValues((val1, val2))
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.res = new_results
+
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.regs = (Region(),)
+
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.successor = [Block(), Block()]
+
+
+def test_op_segmented_accessor_assignment():
+    """Test that operation constructs cannot be assigned via an op's named constructs when using segment size based accessors."""
+
+    @irdl_op_definition
+    class TestSegmentedOp(IRDLOperation):
+        name = "test.segmentedop"
+        operands1 = var_operand_def()
+        operands2 = var_operand_def()
+
+        results1 = var_result_def()
+        results2 = var_result_def()
+
+        regions1 = var_region_def()
+        regions2 = var_region_def()
+
+        successors1 = var_successor_def()
+        successors2 = var_successor_def()
+
+        irdl_options = (
+            AttrSizedOperandSegments(as_property=True),
+            AttrSizedResultSegments(as_property=True),
+            AttrSizedRegionSegments(as_property=True),
+            AttrSizedSuccessorSegments(as_property=True),
+        )
+
+    val1, val2 = create_ssa_value(i32), create_ssa_value(i32)
+    op1 = TestSegmentedOp.build(
+        operands=[[val1], [val2]],
+        result_types=[[], []],
+        regions=[[], []],
+        successors=[[], []],
+    )
+    op1.verify()
+
+    new_operands: SSAValues[SSAValue[Attribute]] = SSAValues((val1, val2))
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.operands1 = new_operands
+
+    new_results: SSAValues[OpResult[Attribute]] = SSAValues((val1, val2))
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.results1 = new_results
+
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.regions1 = (Region(),)
+
+    with pytest.raises(
+        NotImplementedError,
+        match="Cannot write to named operands, regions, results, or successors.",
+    ):
+        op1.successors1 = [Block(), Block()]
 
 
 ################################################################################
